src/main/java/com/APIDevelopmentTask/Todo/app/external/util/JwtTokenUtil.java
package com.APIDevelopmentTask.Todo.app.external.util;

import com.APIDevelopmentTask.Todo.app.application.exception.InvalidTokenException;
import io.jsonwebtoken.Claims;
import io.jsonwebtoken.Jwts;
import io.jsonwebtoken.SignatureAlgorithm;
import lombok.Value;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.stereotype.Component;
import org.springframework.security.core.userdetails.UserDetails;

import java.util.Date;

@Component
public class JwtTokenUtil {
    private static final Logger logger = LoggerFactory.getLogger(JwtTokenUtil.class);

    private String secretKey="secretkey";

    public String generateToken(UserDetails userDetails) {
        String token = Jwts.builder()
                .setSubject(userDetails.getUsername())
                .setExpiration(new Date(System.currentTimeMillis() + 1000 * 60 * 60 * 10))
                .signWith(SignatureAlgorithm.HS512, secretKey)
                .compact();

        logger.info("Generated JWT token for user: {}", userDetails.getUsername());
        return token;
    }

    public String getUsernameFromToken(String token) {
        try {
            Claims claims = Jwts.parser()
                    .setSigningKey(secretKey)
                    .build()
                    .parseClaimsJws(token)
                    .getBody();
            logger.info("Extracted username from token: {}", claims.getSubject());
            return claims.getSubject();
        } catch (Exception e) {
            logger.error("Failed to parse token: {}", token, e);
            throw new InvalidTokenException("Invalid JWT token");
        }

    }
}
